Detect newCopyOnWriteArrayList & newConcurrentMap
Note that the behavior of `Maps.newConcurrentMap()` is slightly
different from `new ConcurrentHashMap<>()`: with Guava 18.0 the former
creates a map with a concurrency level of 4, whereas the latter uses
concurrency level 16. Most users will not care, and those that do
should make this intention explicit.  (This is one of the reasons that
the JavaDoc for Maps.newConcurrentMap points to MapMaker.)