Use process-test-classes phase in README example
There should be no need for the user so sit through the tests before reaching the `validate` stage if the build should fail due to using legacy APIs.

`process-test-classes` makes the most sense because it's the first phase after all classes have been compiled, and works regardless of `includeTestClasses` value.